requirements.txt
pandas==2.0.2
python-dateutil==2.8.2
selenium==4.9.1
webdriver-manager==3.8.6

